fix: safe.bareRepository (#16)
fixes #14

When safe.bareRepository is set to explicit, Git prevents running
commands inside a bare repository unless it is explicitly specified
as the git directory (e.g., via --git-dir). This security feature is
designed to prevent accidental execution of git commands in directories
that happen to look like git repositories
(e.g., a side-effect of CVE-2022-24765).